newest

+++ jstr.h
@@ -1,12 +1,14 @@
 #ifndef JSTR_H_DEF
 #define JSTR_H_DEF
 
+#include <stdlib.h>
+
 #define newJstr(_NAME_OF_STRING, _STRING) jstr _NAME_OF_STRING = {.str = _STRING, .size = 0}
 
 typedef struct Jstr {
 	char *str;
-	int size;
-	int len;
+	size_t size;
+	size_t len;
 } Jstr;
 
 /* end with "\0" or "" since the \0 is implicit */
